server/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
numpy==1.25.1
matplotlib==3.7.2
scipy==1.11.1
shapely==2.0.1
Pillow==9.5.0
opencv-python-headless==4.8.0.74